YG-74 💡 resize image util 함수로 분리
